requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
Pillow
requests
